melhorando nomes e comentando detalhadamente cada verificacao para ativar RTD
