+++ Core/Inc/CAN/inverter_can_monitor.h
@@ -5,10 +5,10 @@
  *      Author: Felipe Telles
  */
 
-#ifndef INC_CAN_INVERTER_CAN_ERROR_MANAGER_H_
-#define INC_CAN_INVERTER_CAN_ERROR_MANAGER_H_
+#ifndef INC_CAN_INVERTER_CAN_MONITOR_H_
+#define INC_CAN_INVERTER_CAN_MONITOR_H_
 
 #define INVERTER_NO_MESSAGE_ERROR_TIME  500
 #define INVERTER_PRECHARGE_TIME         (5000/portTICK_PERIOD_MS)
 
-#endif /* INC_CAN_INVERTER_CAN_ERROR_MANAGER_H_ */
+#endif /* INC_CAN_INVERTER_CAN_MONITOR_H_ */
