+++ Core/Src/RTD.c
@@ -46,6 +46,29 @@ void exit_RTD() {
     osEventFlagsClear(ECU_control_event_id, RTD_FLAG);  //limpa flag de RTD
 }
 
+/*
+ * According to the rules EV.10.4.3 (2022) the only verifications needed to enable RTD are if the brake is active and TS is active.
+ * As an added safety, others verifications are made:
+ *
+ *      - is_brake_active: Allow RTD activation when the brake pedal is being pressed.
+ *
+ *      - !is_throttle_active: Allow RTD activation only if the throttle pedal is not being pressed. Prevents an unexpected
+ *      accelaration of the vehicle, as the driver should not be pressing the pedal before RTD activation
+ *
+ *      - !error_flags &= ALL_SEVERE_ERROR_FLAG: Allow RTD activation if no error that exits RTD is present
+ *
+ *      - race_mode != ERRO: Allow RTD activation after the driver has made an action to ignore the error. The vehicle will
+ *      be on race_mode ERRO after an severe error has exited RTD, the drive must then switch modes again to enable RTD.
+ *
+ *      - is_inverter_ready: Allow RTD activation after the precharge period is over and both AIRs has been closed.
+ *      According to the rules RTD must only be possible after TS is active (which only means the shutdown circuit is closed)
+ *      It also defines RTD as when "the Motors will respond to the input of the APPS". So by the definition the RTD must
+ *      no be possible until both AIRs has been closed.
+ *      This verification should be made reading the status of the AIRs sended by the vehicle BMS, which is not currently
+ *      working. While the BMS cannot make this reading, this verification will enable RTD activation after a timer which
+ *      is started when the inverter sends its first message.
+ *      TODO: Allow RTD activation from the status check of the AIRs.
+ */
 bool can_RTD_be_enabled() {
     uint32_t error_flags = osEventFlagsGet(ECU_control_event_id);                           //obtem todas as flags
     error_flags &= ALL_SEVERE_ERROR_FLAG;                                                   //filtra apenas flags de erros severos, ignorando as outras
@@ -53,7 +76,7 @@ bool can_RTD_be_enabled() {
     THROTTLE_STATUS_t is_throttle_active = get_global_var_value(THROTTLE_STATUS);
     RACE_MODE_t race_mode = get_global_var_value(RACE_MODE);
     bool is_inverter_ready = get_individual_flag(ECU_control_event_id, INVERTER_READY);     // flag that indicates when the inverter precharge time has passed and the inverter is ready
-    if(is_brake_active && !is_throttle_active &&  !error_flags && is_inverter_ready && (race_mode != ERRO))
+    if(is_brake_active && !is_throttle_active &&  !error_flags && (race_mode != ERRO) && is_inverter_ready)
         return true;
     else
         return false;
